chore(medicines): seperated medicine action
